requirements.txt
streamlit
google-generativeai
sentence-transformers
scikit-learn
beautifulsoup4
requests
pdfkit
python-dotenv
pypdf
reportlab
playwright